[OBEY-CAMPAIGN-8deed8b4] fix(workflow): address workflow surface review
